+++ lib/active_currency/rate_store.rb
@@ -3,11 +3,5 @@
 module ActiveCurrency
   class RateStore < DatabaseStore
     include CacheableStore
-
-    attr_reader :currencies
-
-    def initialize(currencies)
-      @currencies = currencies
-    end
   end
 end