+++ spec/dummy/config/initializers/money.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 MoneyRails.configure do |config|
-  rate_store = ActiveCurrency::RateStore.new(%w[EUR USD])
-  config.default_bank = Money::Bank::VariableExchange.new(rate_store)
+  config.default_bank =
+    Money::Bank::VariableExchange.new(ActiveCurrency::RateStore.new)
 end